feat: SemesterDetailScreen ui refactor
